deq annotations refactor
